move current month accordion into its folder of accordions
