+++ src/features/Dashboard/Overview/subscreens/AccountOverviewSubscreen.tsx
@@ -2,7 +2,7 @@ import { Button } from "flowbite-react";
 
 import { AccountDialog } from "@/features/Accounts/AccountDialog";
 import { Account } from "@/features/Accounts/Accounts";
-import { MonthAccordionRecords } from "@/features/Records/CurrentMonthAccordionRecords";
+import { MonthAccordionRecords } from "@/features/Records/Accordions/CurrentMonthAccordionRecords";
 import { useAccountModal } from "@/hooks/useAccountModal";
 import { useDashboard } from "@/shared/hooks/useDashboard";
 import { useDashboardStore } from "@/zustand/provider/dashboard-store-provider";
